profiling: Split scripts to separate impacts of memory-profiler and wss, ensure that debug builds are used

+++ profiling/profiling-wss.sh
@@ -12,8 +12,12 @@ cd "$PROFDIR"
 which actix-web-server || cargo install --path actix-web-server
 which actix-web-server || ( echo "Please add ~/.cargo/bin to your PATH" ; exit 1 )
 which wrk || ( echo "wrk not found: Compile the wrk binary from https://github.com/kinvolk/wrk2/ and move it to your PATH" ; exit 1 )
-ls libmemory_profiler.so memory-profiler-cli || ( curl -L -O https://github.com/nokia/memory-profiler/releases/download/0.3.0/memory-profiler-x86_64-unknown-linux-gnu.tgz ; tar xf memory-profiler-x86_64-unknown-linux-gnu.tgz ; rm memory-profiler-x86_64-unknown-linux-gnu.tgz )
 ls ../target/release/profiling-opt-and-dbg-symbols || ( echo "../target/release/profiling-opt-and-dbg-symbols not found: Please run ./profiling-build.sh" ; exit 1 )
+./wss.pl -h 2> /dev/null && RET=$? || RET=$?
+if [ $RET -eq 2 ]; then
+  echo "wss.pl can not start: Install the perl-Time-HiRes package of your distribution"
+  exit 1
+fi
 
 trap '{ killall iperf actix-web-server >& /dev/null; }' EXIT
 
@@ -30,27 +34,27 @@ single_profiling_run () {
   do
     sleep 1
   done
+  LINKERD_TEST_BIN=$(ls ../target/release/profiling-*[^.d])
+  ( ./wss.pl $(pgrep -f -x "$LINKERD_TEST_BIN --exact profiling_setup --nocapture") 6 | tee "wss_$NAME.$ID.txt" ) &
   if [ "$MODE" = "TCP" ]; then
     iperf -t 6 -p "$PROXY_PORT" -c localhost | tee "$NAME.$ID.txt"
   else
     wrk -L -s wrk-report.lua -R 4500 -H 'Host: transparency.test.svc.cluster.local' "http://127.0.0.1:$PROXY_PORT/" | tee "$NAME.$ID.txt"
   fi
+  while pgrep wss.pl; do
+    sleep 1
+  done
   # signal that proxy can terminate now
   echo F | nc localhost 7777 || true
   # kill server
   kill $SPID
   ) &
-  rm memory-profiling_*.dat || true
-  PROFILING_SUPPORT_SERVER="127.0.0.1:$SERVER_PORT" LD_PRELOAD=./libmemory_profiler.so ../target/release/profiling-opt-and-dbg-symbols --exact profiling_setup --nocapture
-  mv memory-profiling_*.dat "$NAME.$ID.heap.dat"
-  ./memory-profiler-cli export-heaptrack "$NAME.$ID.heap.dat" --output "$NAME.$ID.heaptrack.dat"
+  PROFILING_SUPPORT_SERVER="127.0.0.1:$SERVER_PORT" ../target/release/profiling-opt-and-dbg-symbols --exact profiling_setup --nocapture
 }
 
 MODE=TCP NAME=tcpoutbound PROXY_PORT=$PROXY_PORT_OUTBOUND single_profiling_run
 MODE=TCP NAME=tcpinbound PROXY_PORT=$PROXY_PORT_INBOUND single_profiling_run
 MODE=HTTP NAME=outbound PROXY_PORT=$PROXY_PORT_OUTBOUND single_profiling_run
 MODE=HTTP NAME=inbound PROXY_PORT=$PROXY_PORT_INBOUND single_profiling_run
-ls *$ID*.dat
-echo "a) Run './memory-profiler-cli server CHANGEME.$ID.heap.dat' and open http://localhost:8080/ to browse the memory graphs or,"
-echo "b) run 'heaptrack -a CHANGEME.$ID.heaptrack.dat' to open the heaptrack files for a detailed view."
-echo "(Replace CHANGEME with inbound, outbound, tcpinbound, tcpoutbound.)"
+echo "Inspect memory usage (RSS is main program memory, Ref is actively used memory):"
+ls wss*$ID.txt
